fix: 게시글 조회 수정

+++ src/main/java/com/phcworld/phcworldboardservice/dto/FreeBoardSelectDto.java
@@ -20,7 +20,7 @@ public class FreeBoardSelectDto {
     private LocalDateTime createDate;
     private LocalDateTime updateDate;
     private Integer count;
-    private Long countOfAnswer;
+    private Integer countOfAnswer;
 
     public String getFormattedCreateDate() {
         return LocalDateTimeUtils.getTime(createDate);
